% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprr.R
\name{age}
\alias{age}
\title{Age}
\usage{
age(cpr, date = Sys.Date(), unit = "year")
}
\arguments{
\item{cpr}{Character Vector of CPR numbers with or without hyphens.}

\item{date}{Character or date vector of dates used in computation of ages.
Dates provided as characters must follow the ISO standard, 'yyyy-mm-dd'.}

\item{unit}{Character indicating the unit representing ages. Possible values
are: 'year', 'month', 'week', 'day'.}
}
\value{
A numeric vector of ages. By default, ages are computed in years.
  Note that year and month units are calculated by approximaion, 1 year =
  365.25 days, 1 month = 30.44 days.
}
\description{
Calculate ages from CPR numbers.
}
\examples{
  age(c('1508631111', '1310762222'))

}
