% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprr.R
\name{gender}
\alias{gender}
\title{Gender}
\usage{
gender(cpr)
}
\arguments{
\item{cpr}{Character Vector of CPR numbers with or without hyphens.}
}
\value{
Integer vector of genders. 0 = Female, 1 = Male.
}
\description{
Calculate genders from CPR numbers.
}
\examples{
  gender(c('1508631111', '1310762222'))

}
