% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprr.R
\name{mod11}
\alias{mod11}
\title{Modulo 11 check}
\usage{
mod11(cpr)
}
\arguments{
\item{cpr}{Character vector of CPR numbers with or without hyphens.}
}
\value{
Logical vector of check results, NA if CPR number includes characters
  (temporary CPR number).
}
\description{
Check if CPR numbers conform to modulo 11 check. Note, modulo 11 check was
deprecated in 2007.
}
\examples{
  mod11(c('1508631111', '1310762222', '2110625629'))

}
