% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_cpr}
\alias{test_cpr}
\title{CPR numbers for testing.}
\format{A data frame with 30 rows and 3 columns.
\describe{
  \item{cpr}{CPR number}
  \item{dob}{Date of birth}
  \item{gender}{Gender, 0 = female, 1 = male}
  ...
}}
\source{
\url{http://medcom.dk/opslag/koder-tabeller-ydere/tabeller/nationale-test-cpr-numre}
}
\usage{
test_cpr
}
\description{
A dataset containing the list of CPR numbers used for testing calculation
of date of birth and gender.
}
\keyword{datasets}
