% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprr.R
\name{dob}
\alias{dob}
\title{Date of birth}
\usage{
dob(cpr)
}
\arguments{
\item{cpr}{Character Vector of CPR numbers with or without hyphens.}
}
\value{
Vector of dates of birth.
}
\description{
Calculate dates of birth from CPR numbers.
}
\examples{
  dob(c('1508631111', '1310762222'))

}
