% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprr.R
\name{scramble}
\alias{scramble}
\title{Scramble}
\usage{
scramble(cpr)
}
\arguments{
\item{cpr}{Character Vector of CPR numbers with or without hyphens.}
}
\value{
A data frame with corresponding CPR and anonymous id numbers.
}
\description{
Make "ananymous" random CPR numbers to random numbers.
}
\examples{
  scramble(c('1508631111', '1310762222'))

}
