% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cprr.R
\name{scramble}
\alias{scramble}
\title{Scramble}
\usage{
scramble(cpr)
}
\arguments{
\item{cpr}{Character vector of CPR numbers with or without hyphens.}
}
\value{
Integer vector. Each integer represents one unique CPR number.
}
\description{
Make "anonymous" random numbers from CPR numbers.
}
\examples{
  scramble(c('1508631111', '1310762222', '1508631111'))

}
